Added float for extra feature

+++ src/components/BoughtCoincard.js
@@ -5,8 +5,8 @@ export default function BoughtCoincard({ coin, currentTheme, coinData }) {
     return (
         <tr>
             <td>{coin.coin}</td>
-            <td>{parseInt(coin.amount)}</td>
-            {parseInt(coin.amount) > 0 && <td onClick={() => { navigate("/trades/" + coin.coin + ":sell") }} className="sell" style={currentTheme}>Sell</td>}
+            <td>{parseFloat(coin.amount)}</td>
+            {parseFloat(coin.amount) > 0 && <td onClick={() => { navigate("/trades/" + coin.coin + ":sell") }} className="sell" style={currentTheme}>Sell</td>}
             <td onClick={() => { navigate("/buy-crypto/" + coinData.name) }} className="buy" style={currentTheme}>Buy</td>
         </tr >
     )
